Move the testing domain's genesis config to the testing consensus chain's genesis config
